requirements.txt
langchain
langchain-openai
langchain-community
pydantic
pandas
openai
sacrebleu
unbabel-comet
python-dotenv
sentence-transformers
faiss-cpu
